requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.28.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.109.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
orjson==3.10.7
passlib==1.7.4
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
psycopg2==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23..
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.39
starlette==0.35.1
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
